requirements.txt
backcall==0.1.0
decorator==4.3.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
numpy==1.15.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
